refactor: transform color module
