+++ src/modules/color/_types.ts
@@ -1,0 +1,3 @@
+export type StringColorFormat = 'css' | 'binary';
+export type NumberColorFormat = 'decimal';
+export type ColorFormat = StringColorFormat | NumberColorFormat;
