+++ src/modules/color/css-supported-function.ts
@@ -2,6 +2,26 @@ import type { FakerCore } from '../../core';
 import { enumValue } from '../helpers/enum-value';
 
 /**
+ * Functions supported by CSS to produce color.
+ */
+export enum CssFunction {
+  RGB = 'rgb',
+  RGBA = 'rgba',
+  HSL = 'hsl',
+  HSLA = 'hsla',
+  HWB = 'hwb',
+  CMYK = 'cmyk',
+  LAB = 'lab',
+  LCH = 'lch',
+  COLOR = 'color',
+}
+
+/**
+ * Functions supported by CSS to produce color.
+ */
+export type CssFunctionType = `${CssFunction}`;
+
+/**
  * Returns a random CSS-supported color function name.
  *
  * @param fakerCore The FakerCore to use.
