+++ src/modules/color/index.ts
@@ -1,1 +1,10 @@
+export type {
+  ColorFormat,
+  NumberColorFormat,
+  StringColorFormat,
+} from './_types';
+export { CssFunction } from './css-supported-function';
+export type { CssFunctionType } from './css-supported-function';
+export { CssSpace } from './css-supported-space';
+export type { CssSpaceType } from './css-supported-space';
 export * from './module';
