+++ src/modules/color/lch.ts
@@ -1,5 +1,11 @@
 import type { FakerCore } from '../../core';
 import { float } from '../number/float';
+import { toColorFormat } from './_to-color-format';
+import type {
+  ColorFormat,
+  NumberColorFormat,
+  StringColorFormat,
+} from './_types';
 
 /**
  * Returns an LCH color. Even though upper bound of
