+++ src/modules/color/module.ts
@@ -1,202 +1,16 @@
 import { ModuleBase } from '../../internal/module-base';
 import type { Casing } from '../../utils/types';
-
-/**
- * Color space names supported by CSS.
- */
-export enum CssSpace {
-  SRGB = 'sRGB',
-  DisplayP3 = 'display-p3',
-  REC2020 = 'rec2020',
-  A98RGB = 'a98-rgb',
-  ProphotoRGB = 'prophoto-rgb',
-}
-
-/**
- * Color space names supported by CSS.
- */
-export type CssSpaceType = `${CssSpace}`;
-
-/**
- * Functions supported by CSS to produce color.
- */
-export enum CssFunction {
-  RGB = 'rgb',
-  RGBA = 'rgba',
-  HSL = 'hsl',
-  HSLA = 'hsla',
-  HWB = 'hwb',
-  CMYK = 'cmyk',
-  LAB = 'lab',
-  LCH = 'lch',
-  COLOR = 'color',
-}
-
-/**
- * Functions supported by CSS to produce color.
- */
-export type CssFunctionType = `${CssFunction}`;
-
-export type StringColorFormat = 'css' | 'binary';
-export type NumberColorFormat = 'decimal';
-export type ColorFormat = StringColorFormat | NumberColorFormat;
-
-/**
- * Formats the hex format of a generated color string according
- * to options specified by user.
- *
- * @param hexColor Hex color string to be formatted.
- * @param options Options object.
- * @param options.prefix Prefix of the generated hex color.
- * @param options.casing Letter type case of the generated hex color.
- */
-function formatHexColor(
-  hexColor: string,
-  options: {
-    prefix: string;
-    casing: Casing;
-  }
-): string {
-  const { prefix, casing } = options;
-
-  switch (casing) {
-    case 'upper': {
-      hexColor = hexColor.toUpperCase();
-      break;
-    }
-
-    case 'lower': {
-      hexColor = hexColor.toLowerCase();
-      break;
-    }
-
-    case 'mixed':
-    // Do nothing
-  }
-
-  if (prefix) {
-    hexColor = prefix + hexColor;
-  }
-
-  return hexColor;
-}
-
-/**
- * Converts an array of numbers into binary string format.
- *
- * @param values Array of values to be converted.
- */
-function toBinary(values: number[]): string {
-  const binary: string[] = values.map((value) => {
-    const isFloat = value % 1 !== 0;
-    if (isFloat) {
-      const buffer = new ArrayBuffer(4);
-      new DataView(buffer).setFloat32(0, value);
-      const bytes = new Uint8Array(buffer);
-      return toBinary([...bytes]).replaceAll(' ', '');
-    }
-
-    return (value >>> 0).toString(2).padStart(8, '0');
-  });
-  return binary.join(' ');
-}
-
-/**
- * Converts the given value to a percentage (`round(value * 100)`).
- *
- * @param value The value to convert to a percentage.
- */
-function toPercentage(value: number): number {
-  return Math.round(value * 100);
-}
-
-/**
- * Converts an array of numbers into CSS accepted format.
- *
- * @param values Array of values to be converted.
- * @param cssFunction CSS function to be generated for the color. Defaults to `'rgb'`.
- * @param space Color space to format CSS color function with. Defaults to `'sRGB'`.
- */
-function toCSS(
-  values: number[],
-  cssFunction: CssFunctionType = 'rgb',
-  space: CssSpaceType = 'sRGB'
-): string {
-  switch (cssFunction) {
-    case 'rgba': {
-      return `rgba(${values[0]}, ${values[1]}, ${values[2]}, ${values[3]})`;
-    }
-
-    case 'color': {
-      return `color(${space} ${values[0]} ${values[1]} ${values[2]})`;
-    }
-
-    case 'cmyk': {
-      return `cmyk(${toPercentage(values[0])}%, ${toPercentage(
-        values[1]
-      )}%, ${toPercentage(values[2])}%, ${toPercentage(values[3])}%)`;
-    }
-
-    case 'hsl': {
-      return `hsl(${values[0]}deg ${toPercentage(values[1])}% ${toPercentage(
-        values[2]
-      )}%)`;
-    }
-
-    case 'hsla': {
-      return `hsl(${values[0]}deg ${toPercentage(values[1])}% ${toPercentage(
-        values[2]
-      )}% / ${values[3]})`;
-    }
-
-    case 'hwb': {
-      return `hwb(${values[0]} ${toPercentage(values[1])}% ${toPercentage(
-        values[2]
-      )}%)`;
-    }
-
-    case 'lab': {
-      return `lab(${toPercentage(values[0])}% ${values[1]} ${values[2]})`;
-    }
-
-    case 'lch': {
-      return `lch(${toPercentage(values[0])}% ${values[1]} ${values[2]})`;
-    }
-
-    case 'rgb': {
-      return `rgb(${values[0]}, ${values[1]}, ${values[2]})`;
-    }
-  }
-}
-
-/**
- * Converts an array of color values to the specified color format.
- *
- * @param values Array of color values to be converted.
- * @param format Format of generated RGB color.
- * @param cssFunction CSS function to be generated for the color. Defaults to `'rgb'`.
- * @param space Color space to format CSS color function with. Defaults to `'sRGB'`.
- */
-function toColorFormat(
-  values: number[],
-  format: ColorFormat,
-  cssFunction: CssFunctionType = 'rgb',
-  space: CssSpaceType = 'sRGB'
-): string | number[] {
-  switch (format) {
-    case 'css': {
-      return toCSS(values, cssFunction, space);
-    }
-
-    case 'binary': {
-      return toBinary(values);
-    }
-
-    case 'decimal': {
-      return values;
-    }
-  }
-}
+import { formatHexColor } from './_format-hex-color';
+import { toColorFormat } from './_to-color-format';
+import type {
+  ColorFormat,
+  NumberColorFormat,
+  StringColorFormat,
+} from './_types';
+import type { CssFunctionType } from './css-supported-function';
+import { CssFunction } from './css-supported-function';
+import type { CssSpaceType } from './css-supported-space';
+import { CssSpace } from './css-supported-space';
 
 /**
  * Module to generate colors.
