+++ src/modules/color/rgb.ts
@@ -3,6 +3,14 @@ import type { Casing } from '../../utils/types';
 import { float } from '../number/float';
 import { int } from '../number/int';
 import { hexadecimal } from '../string/hexadecimal';
+import { formatHexColor } from './_format-hex-color';
+import { toColorFormat } from './_to-color-format';
+import type {
+  ColorFormat,
+  NumberColorFormat,
+  StringColorFormat,
+} from './_types';
+import type { CssFunctionType } from './css-supported-function';
 
 /**
  * Returns an RGB color.
